Job Services API call refactoring
