+++ src/jobs/jobDetails.tsx
@@ -26,15 +26,10 @@ import EditIcon from '../../style/icons/edit_icon.svg';
 import EditIconDisable from '../../style/icons/edit_icon_disable.svg';
 import DeletePopup from '../utils/deletePopup';
 import {
-  API_HEADER_BEARER,
-  API_HEADER_CONTENT_TYPE,
-  BASE_URL,
   JOB_FIELDS_EXCLUDED,
-  LABEL_TEXT,
   STATUS_RUNNING
 } from '../utils/const';
 import {
-  authApi,
   elapsedTime,
   jobDetailsOptionalDisplay,
   jobTimeFormat,
@@ -42,8 +37,7 @@ import {
   jobTypeValue,
   jobTypeValueArguments,
   statusMessage,
-  toastifyCustomStyle,
-  loggedFetch
+  toastifyCustomStyle
 } from '../utils/utils';
 
 import ClusterDetails from '../cluster/clusterDetails';
@@ -58,7 +52,6 @@ import { JobService } from './jobServices';
 import errorIcon from '../../style/icons/error_icon.svg';
 import PollingTimer from '../utils/pollingTimer';
 import { IJobDetails } from '../utils/jobDetailsInterface';
-import { DataprocLoggingService, LOG_LEVEL } from '../utils/loggingService';
 
 const iconLeftArrow = new LabIcon({
   name: 'launcher:left-arrow-icon',
@@ -192,7 +185,7 @@ function JobDetails({
 
   const handleDetailedClusterView = () => {
     pollingJobDetails(getJobDetails, true);
-    if (fromPage !== "clusters") {
+    if (fromPage !== 'clusters') {
       setDetailedJobView(false);
     }
     setDetailedClusterView(true);
@@ -233,47 +226,7 @@ function JobDetails({
   };
 
   const updateJobDetails = async (payloadJob: object) => {
-    const credentials = await authApi();
-    if (credentials) {
-      loggedFetch(
-        `${BASE_URL}/projects/${credentials.project_id}/regions/${credentials.region_id}/jobs/${jobSelected}?updateMask=${LABEL_TEXT}`,
-        {
-          method: 'PATCH',
-          body: JSON.stringify(payloadJob),
-          headers: {
-            'Content-Type': API_HEADER_CONTENT_TYPE,
-            Authorization: API_HEADER_BEARER + credentials.access_token
-          }
-        }
-      )
-        .then((response: Response) => {
-          response
-            .json()
-            .then(async (responseResultJob: Response) => {
-              const formattedResponse = await responseResultJob.json()
-              if (formattedResponse?.error?.code) {
-                toast.error(formattedResponse?.error?.message, toastifyCustomStyle);
-              }
-            
-             
-              console.log(responseResultJob);
-              toast.success(
-                `Request to update job ${jobSelected} submitted`,
-                toastifyCustomStyle
-              );
-              
-            })
-            .catch((e: Error) => console.error(e));
-        })
-        .catch((err: Error) => {
-          console.error('Error in updating job', err);
-          DataprocLoggingService.log('Error in updating job', LOG_LEVEL.ERROR);
-          toast.error(
-            `Failed to update the job ${jobSelected}`,
-            toastifyCustomStyle
-          );
-        });
-    }
+    await JobService.updateJobDetailsService(payloadJob, jobSelected);
   };
 
   const handleSaveEdit = () => {
@@ -305,73 +258,35 @@ function JobDetails({
   };
 
   const getJobDetails = async () => {
-    const credentials = await authApi();
-    if (credentials) {
-      loggedFetch(
-        `${BASE_URL}/projects/${credentials.project_id}/regions/${credentials.region_id}/jobs/${jobSelected}`,
-        {
-          method: 'GET',
-          headers: {
-            'Content-Type': API_HEADER_CONTENT_TYPE,
-            Authorization: API_HEADER_BEARER + credentials.access_token
-          }
-        }
-      )
-        .then((response: Response) => {
-          response
-            .json()
-            .then((responseResult: any) => {
-              if (responseResult.error && responseResult.error.code === 404) {
-                setErrorView(true);
-              }
-              setjobInfoResponse(responseResult);
-              const labelValue: string[] = [];
-              if (responseResult.labels) {
-                for (const [key, value] of Object.entries(
-                  responseResult.labels
-                )) {
-                  labelValue.push(`${key}:${value}`);
-                }
-              }
-              setjobInfo(responseResult);
-              setLabelDetail(labelValue);
-              setIsLoading(false);
-              setSelectedJobClone(responseResult);
-              if (responseResult?.error?.code) {
-                toast.error(responseResult?.error?.message, toastifyCustomStyle);
-              }
-            })
-            .catch((e: Error) => {
-              console.error(e);
-              setIsLoading(false);
-            });
-        })
-        .catch((err: Error) => {
-          setIsLoading(false);
-          console.error('Error in getting job details', err);
-          DataprocLoggingService.log('Error in getting job details', LOG_LEVEL.ERROR);
-          toast.error(
-            `Failed to fetch job details ${jobSelected}`,
-            toastifyCustomStyle
-          );
-        });
-    }
+    await JobService.getJobDetailsService(
+      jobSelected,
+      setErrorView,
+      setIsLoading,
+      setjobInfoResponse,
+      setjobInfo,
+      setLabelDetail,
+      setSelectedJobClone
+    );
   };
 
-  const startTime = !errorView?jobTimeFormat(jobInfo.statusHistory[0].stateStartTime):"";
-  const job =!errorView? jobTypeValue(jobInfo):"";
-  const jobType = !errorView?jobTypeDisplay(job):"";
-  const jobArgument = !errorView?jobTypeValueArguments(jobInfo):"";
-  const jobTypeConcat = !errorView?jobArgument + 'Job':"";
+  const startTime = !errorView
+    ? jobTimeFormat(jobInfo.statusHistory[0].stateStartTime)
+    : '';
+  const job = !errorView ? jobTypeValue(jobInfo) : '';
+  const jobType = !errorView ? jobTypeDisplay(job) : '';
+  const jobArgument = !errorView ? jobTypeValueArguments(jobInfo) : '';
+  const jobTypeConcat = !errorView ? jobArgument + 'Job' : '';
   //@ts-ignore string used as index
-  const argumentsList =!errorView? jobInfo[jobTypeConcat].args:"";
-  const statusMsg = !errorView?statusMessage(jobInfo):"";
-  const endTime =!errorView?new Date(jobInfo.status.stateStartTime):new Date;
-  const jobStartTime =!errorView? new Date(
-    jobInfo.statusHistory[0].stateStartTime
-  ):new Date;
+  const argumentsList = !errorView ? jobInfo[jobTypeConcat].args : '';
+  const statusMsg = !errorView ? statusMessage(jobInfo) : '';
+  const endTime = !errorView
+    ? new Date(jobInfo.status.stateStartTime)
+    : new Date();
+  const jobStartTime = !errorView
+    ? new Date(jobInfo.statusHistory[0].stateStartTime)
+    : new Date();
 
-  let elapsedTimeString = !errorView?elapsedTime(endTime, jobStartTime):"";
+  let elapsedTimeString = !errorView ? elapsedTime(endTime, jobStartTime) : '';
   const statusStyleSelection = (jobInfo: any) => {
     if (jobInfo.status.state === STATUS_RUNNING) {
       return 'action-cluster-section'; //CSS class
@@ -388,7 +303,7 @@ function JobDetails({
     if (labelEditMode) {
       return 'job-edit-button-disabled';
     } else {
-      return  'job-edit-button';
+      return 'job-edit-button';
     }
   };
 
@@ -558,19 +473,18 @@ function JobDetails({
                   {labelEditMode ? (
                     <iconEditDisable.react
                       tag="div"
-                      className= {styleIconColor(labelEditMode)}
+                      className={styleIconColor(labelEditMode)}
                     />
                   ) : (
                     <iconEdit.react
                       tag="div"
-                      className={styleIconColor(labelEditMode)
-                      }
+                      className={styleIconColor(labelEditMode)}
                     />
                   )}
                   <div
-                    className={labelEditMode
-                      ? 'job-edit-text-disabled'
-                      : 'job-edit-text'}
+                    className={
+                      labelEditMode ? 'job-edit-text-disabled' : 'job-edit-text'
+                    }
                   >
                     EDIT
                   </div>
@@ -760,9 +674,7 @@ function JobDetails({
                 </div>
                 {labelEditMode && (
                   <div className="job-button-style-parent">
-                    <div
-                      className='job-save-button-style'
-                    >
+                    <div className="job-save-button-style">
                       <div
                         role="button"
                         onClick={() => {
@@ -789,22 +701,20 @@ function JobDetails({
           )}
 
           {jobInfo.jobUuid === '' && (
-           <>
-           {isLoading && (
-         <div className="spin-loaderMain">
-          <ClipLoader
-              color="#3367d6"
-             loading={true}
-             size={18}
-             aria-label="Loading Spinner"
-             data-testid="loader"
-           />
-           Loading Job Details
-         </div>
-       )}
-        
-       </>
-     
+            <>
+              {isLoading && (
+                <div className="spin-loader-main">
+                  <ClipLoader
+                    color="#3367d6"
+                    loading={true}
+                    size={18}
+                    aria-label="Loading Spinner"
+                    data-testid="loader"
+                  />
+                  Loading Job Details
+                </div>
+              )}
+            </>
           )}
         </div>
       )}
